[CK_TILE] Refactor GemmKernel - review changes
